This is a file with comments
This is to a file that lacked comments from the people who kindly handed ePaper display code. It made me waste so much time in trying to understand it. So I decided to write a few comments down for when I hadn't understood its initial purpose. Their are some parts that lack clear indications of what "that" part of the code "really" does. But, that is simply because I hadn't had the need to inheritance understand it. If it was my job to correctly pindown what everything did, I would do it without hesitation. Since this is more of a note taking than actual comments. I put it here to potentially guide someone what went wrong.